+++ verify_setup.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+# -*- coding: ascii -*-
 from datetime import date
 from _version import __version__
 import requests